refactor(m3): widen gate_literal_prefix_size to int64_t
Per Gemini's review on PR #80: align the runner-side flag type with
the comparator API (`CompareLiteralToWtns(... int64_t prefix_size)`)
and the natural type of `num_elements` (int64_t, line 202). Removes
the implicit int32→int64 widen at line 203 and avoids tsl::Flag parse
failure on hypothetical >INT32_MAX prefix values for circuits whose
output exceeds 2B elements.

`tsl::Flag` has an `int64_t*` constructor (xla/tsl/util/command_line_flags.h:69),
so the binding remains valid. Behavior diff = 0 for current consumers.

Test: //bench/m3:witness_compare_test PASSED.
      //bench/m3:m3_correctness_gate_test PASSED 36.2s (26/26).

+++ bench/m3/m3_runner_main.cc
@@ -90,7 +90,7 @@ struct Options {
   // N indices and N must be ≤ literal element count. Used for chips whose
   // GPU output flattens public outputs followed by yet-to-be-aligned internal
   // signals (canonical: `aes_256_encrypt`, `tensor<14048>`, prefix=128).
-  int32_t gate_literal_prefix_size = 0;
+  int64_t gate_literal_prefix_size = 0;
 };
 
 absl::StatusOr<std::vector<zkx::Literal>>
